🏗 Restructure CL2.0 routes
