+++ src/components/SupportingLikerView/SupportingLikerView.vue
@@ -13,7 +13,7 @@
         {{ `${price} ${$t('Currency.USD')}/${$t('SubscriptionPeriod.Month')}` }}</div>
       <nuxt-link
         class="mt-8 text-12 text-like-green underline"
-        :to="{ name: 'settings-support-users-id', params: { id: likerId } }"
+        :to="{ name: 'id-civic', params: { id: likerId } }"
       >{{ $t('edit') }}</nuxt-link>
     </template>
   </PureSupportingLikerView>
