+++ src/pages/settings.vue
@@ -5,10 +5,7 @@
         SiteNavBar.text-like-green
 
     main.page-content.page-content--narrow
-      SettingsPageHeader(
-        :class="headerClass"
-        :is-show-back="isHeaderShowBack"
-      )
+      SettingsPageHeader(:is-show-back="isHeaderShowBack")
 
       NuxtChild
 </template>
@@ -26,12 +23,6 @@ export default {
     SettingsPageHeader,
   },
   computed: {
-    headerClass() {
-      const isHideHeader = this.$route.name === 'settings-support-users-id';
-      return {
-        'opacity-0 pointer-events-none': isHideHeader,
-      };
-    },
     isHeaderShowBack() {
       return /^settings-.+$/.test(this.$route.name);
     },
