--- src/pages/settings/support/users/index.vue
@@ -1,10 +1,0 @@
-<script>
-export default {
-  fetch({ redirect, query }) {
-    redirect(302, {
-      name: 'settings-support',
-      query,
-    });
-  },
-};
-</script>
