+++ src/util/billing.js
@@ -1,0 +1,15 @@
+export function getMaskedCardNumber(brand, last4) {
+  switch (brand) {
+    case 'visa':
+    case 'mastercard':
+      return `•••• •••• •••• ${last4}`;
+
+    case 'amex':
+      return `•••• •••••• •${last4}`;
+
+    default:
+      return `•••• ${last4}`;
+  }
+}
+
+export default { getMaskedCardNumber };
